+++ src/Core/Json/Hash.php
@@ -62,7 +62,6 @@ class Hash implements ArrayAccess, Arrayable, Countable, IteratorAggregate, Json
     public function __construct(array $value = null)
     {
         $this->value = $value ?: [];
-        $this->sorted();
     }
 
     /**