+++ src/Core/Schema/Schema.php
@@ -44,9 +44,26 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
     protected int $maxDepth = 1;
 
     /**
+     * The key name for the resource "id".
+     *
+     * @var string|null
+     */
+    protected ?string $idKeyName = null;
+
+    /**
      * @var array|null
      */
     private ?array $fields = null;
+
+    /**
+     * @var array|null
+     */
+    private ?array $attributes = null;
+
+    /**
+     * @var array|null
+     */
+    private ?array $relations = null;
 
     /**
      * @var callable|null
@@ -158,6 +175,18 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
     /**
      * @inheritDoc
      */
+    public function idKeyName(): ?string
+    {
+        if ($this->idKeyName) {
+            return $this->idKeyName;
+        }
+
+        return $this->id()->key();
+    }
+
+    /**
+     * @inheritDoc
+     */
     public function fieldNames(): array
     {
         return array_keys($this->allFields());
@@ -192,11 +221,7 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
      */
     public function attributes(): iterable
     {
-        foreach ($this as $field) {
-            if ($field instanceof Attribute) {
-                yield $field->name() => $field;
-            }
-        }
+        yield from $this->allAttributes();
     }
 
     /**
@@ -204,9 +229,7 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
      */
     public function attribute(string $name): Attribute
     {
-        $field = $this->allFields()[$name] ?? null;
-
-        if ($field instanceof Attribute) {
+        if ($field = $this->allAttributes()[$name] ?? null) {
             return $field;
         }
 
@@ -222,7 +245,7 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
      */
     public function isAttribute(string $name): bool
     {
-        $field = $this->allFields()[$name] ?? null;
+        $field = $this->allAttributes()[$name] ?? null;
 
         return $field instanceof Attribute;
     }
@@ -232,11 +255,7 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
      */
     public function relationships(): iterable
     {
-        foreach ($this as $field) {
-            if ($field instanceof Relation) {
-                yield $field->name() => $field;
-            }
-        }
+        yield from $this->allRelations();
     }
 
     /**
@@ -244,9 +263,7 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
      */
     public function relationship(string $name): Relation
     {
-        $field = $this->allFields()[$name] ?? null;
-
-        if ($field instanceof Relation) {
+        if ($field = $this->allRelations()[$name] ?? null) {
             return $field;
         }
 
@@ -262,7 +279,7 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
      */
     public function isRelationship(string $name): bool
     {
-        $field = $this->allFields()[$name] ?? null;
+        $field = $this->allRelations()[$name] ?? null;
 
         return $field instanceof Relation;
     }
@@ -337,5 +354,33 @@ abstract class Schema implements SchemaContract, SchemaAwareContract, \IteratorA
 
             return $field->name();
         })->sortKeys()->all();
+    }
+
+    /**
+     * @return array
+     */
+    private function allAttributes(): array
+    {
+        if (is_array($this->attributes)) {
+            return $this->attributes;
+        }
+
+        return $this->attributes = collect($this->allFields())
+            ->whereInstanceOf(Attribute::class)
+            ->all();
+    }
+
+    /**
+     * @return array
+     */
+    private function allRelations(): array
+    {
+        if (is_array($this->relations)) {
+            return $this->relations;
+        }
+
+        return $this->relations = collect($this->allFields())
+            ->whereInstanceOf(Relation::class)
+            ->all();
     }
 }